Refactor data preprocessing (#53)
A function that takes an AusTraits database and creates all the relevant
files needed for the portal, with standardised name in a nominated
directory. This means swapping datasets (e.g. lite, mid, full), only
requires swapping out a single path -the folder name, and minimises
repeated code in data preprocessing.

---------

Co-authored-by: Fonti Kar <f.kar@unsw.edu.au>